requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-telegram-bot==20.7
python-multipart
aiofiles
httpx
jinja2
python-dotenv
openai